Comment on message marshalling, and added session id
Added a comment stating that, while the message passing syntax is
really verbose in the draft, a real implementation could use something
like protobufs to limit the number of bytes on the wire.

I also added a session_id part to the connection confirmation message.
This session is must then somehow be passed from a client to a device
upon connecting. It can also be used by the personal signpost to tell a
device that an open connection should be terminated because of changes
in policy.