Dummy code for setting up a client <--> personal signpost connection
This sets up a dummy connection between a signpost client and a
personal signpost.
It doesn't yet download client public keys it doesn't have, nor does it
validate the validity of any keys. That is yet to be done.
The ticket_id assigned is also static at the moment, and needs to be
stored somewhere.