Dummy tactic solver
Got crazy not doing anything real, so spent a few hours coding up a
data flow tactic solver.
If this is something we end up using or not, is not too important, but
it was an interesting experience regardless.

Tactics can be added to the tactics folder, and will automatically be
used.