Add TTLs to link evaluations
TTLs are added to link evaluation (strategy instances). Nodes remove
links as they time out. The node responsible for the link reevaluates
the strategy before the link is invalidated. To ensure that a link
isn't continuously removed and then readied, the node responsible for
the link takes into account how long it took to previously evaluate the
link when deciding when to reevaluate it.