Increased sleep parameter in unit test
Tests that rely on bud doing it's work unfortunately at the moment
rely on a timer to wait long enough for the work to complete.
This is not a good solution as bud needs shorter or longer depending
on the hardware the tests are run on.
Increased the timeout somewhat to have tests pass on slower dev
machines