[NOT BACKWARDS COMPATIBLE] Comm using unix sockets
This commit is not backwards compatible!
The tactics are now communicating with the
tactics scripts using unix sockets rather than stdio.
The unix socket is passed to the tactic script as the
first and only argument.

The tactics now use EventMachine connections rather
than threads. This should be cleaner, and hopefully better.

Performance is still abysmal.