Implement recycling tactics and tactic thread pool
This was a necessary change to get around the
tactics solvers abysmal performance.
Tactic instances now live in a tactic thread pool
which lazily expands as required.
Instead of terminating, a tactic should now instead
recycle itself (all green and sustainable!).
It does so by sending the tactic solver a

    {"recycle":true}

message, and resetting itself to its initial state.