Enable broadcast of resolve requests and truths
Now, imagine we have three nodes A, B and C.
A and C both see B, but not each other.
Now, if A wants to resolve something on node C,
it requires B to relay this information to C.
Therefore we now have an ugly broadcast, where
everyone broadcasts whatever they get to everyone...
O(n^2) in the case where everyone is connected.
Not nice, but will work for our tests.