Add new dummy engine
This is engine number one trillion.
What it currently does, is create a static
list of all the possible link combinations
we can create. Then later on, when we need
a link of a certain characteristic, we can
filter the list to find the possible
combos that produce at least the required
characteristics.